+++ CHANGELOG.md
@@ -5,10 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.10.0] - 2021-05-19
+[Unreleased]
+
+### Added
+
 
 ### Changed
-- Update guide for priority based downlink policy
+- Applying matrix3d(-1, 0, 0, 0, 0, 1, 0, 0, 0, 0, -1, 0, 0, 0, 0, 1) in stead of rotateY(180deg) to local videoTile in order to fix blank video tile in Safari.
+
+### Removed
+
+
+### Fixed
+
+
+## [2.10.0] - 2021-05-19
 
 ### Added 
 
@@ -24,6 +35,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Bump version for lodash, y18n, and ssri dependencies.
 - Mark `getObservableVideoMetrics` optional in ClientMetricReprt and `videoStreamIndex` and `selfAttendeeId` optional in `DefaultClientMetricReport`.
+- Update guide for priority based downlink policy
+
 
 ### Removed